@@ -12,7 +12,7 @@ import re
 import unicodedata
 from dataclasses import dataclass, field
 from datetime import datetime
-from typing import Any, Optional
+from typing import Any
 
 
 @dataclass
@@ -31,6 +31,9 @@ class Contact:
         organizations: List of organization names
         notes: Contact notes
         last_modified: Timestamp of last modification
+        photo_url: URL to contact's photo
+        photo_data: Binary photo data
+        photo_etag: ETag for photo version tracking
 
     Usage:
         # Create from API response
@@ -50,13 +53,19 @@ class Contact:
     etag: str  # Required for updates
     display_name: str
 
-    given_name: Optional[str] = None
-    family_name: Optional[str] = None
+    given_name: str | None = None
+    family_name: str | None = None
     emails: list[str] = field(default_factory=list)
     phones: list[str] = field(default_factory=list)
     organizations: list[str] = field(default_factory=list)
-    notes: Optional[str] = None
-    last_modified: Optional[datetime] = None
+    notes: str | None = None
+    last_modified: datetime | None = None
+    memberships: list[str] = field(default_factory=list)  # Contact group resource names
+
+    # Photo fields
+    photo_url: str | None = None
+    photo_data: bytes | None = None
+    photo_etag: str | None = None
 
     # Additional fields for sync tracking
     deleted: bool = False  # True if contact was deleted in source
@@ -119,6 +128,13 @@ class Contact:
         biographies = person.get("biographies", [])
         notes = biographies[0].get("value") if biographies else None
 
+        # Extract memberships (contact group resource names)
+        memberships = [
+            m.get("contactGroupMembership", {}).get("contactGroupResourceName", "")
+            for m in person.get("memberships", [])
+            if m.get("contactGroupMembership", {}).get("contactGroupResourceName")
+        ]
+
         # Extract last modified time from metadata
         last_modified = None
         metadata = person.get("metadata", {})
@@ -137,6 +153,19 @@ class Contact:
         # Check if contact is deleted
         deleted = person.get("metadata", {}).get("deleted", False)
 
+        # Extract photo information
+        photo_url = None
+        photos = person.get("photos", [])
+        if photos:
+            # Look for primary photo first
+            primary_photo = next(
+                (p for p in photos if p.get("metadata", {}).get("primary", False)),
+                None,
+            )
+            # Use primary photo if found, otherwise use first photo
+            photo = primary_photo if primary_photo else photos[0]
+            photo_url = photo.get("url")
+
         return cls(
             resource_name=person.get("resourceName", ""),
             etag=person.get("etag", ""),
@@ -148,6 +177,8 @@ class Contact:
             organizations=organizations,
             notes=notes,
             last_modified=last_modified,
+            memberships=memberships,
+            photo_url=photo_url,
             deleted=deleted,
         )
 
@@ -162,6 +193,8 @@ class Contact:
             - Does not include resourceName (set by Google on create)
             - Does not include etag (should be passed separately for updates)
             - Only includes non-empty fields
+            - Photos are included for informational purposes but must be
+              updated separately via the updateContactPhoto endpoint
         """
         person: dict[str, Any] = {}
 
@@ -193,6 +226,18 @@ class Contact:
         # Add notes as biography
         if self.notes:
             person["biographies"] = [{"value": self.notes, "contentType": "TEXT_PLAIN"}]
+
+        # Add memberships (contact group assignments)
+        if self.memberships:
+            person["memberships"] = [
+                {"contactGroupMembership": {"contactGroupResourceName": m}}
+                for m in self.memberships
+            ]
+
+        # Add photo information
+        # Note: Photos are updated separately via updateContactPhoto endpoint
+        if self.photo_url:
+            person["photos"] = [{"url": self.photo_url, "metadata": {"primary": True}}]
 
         return person
 
@@ -320,6 +365,8 @@ class Contact:
             f"phones:{','.join(sorted(self._normalize_phones()))}",
             f"organizations:{','.join(sorted(self.organizations))}",
             f"notes:{self.notes or ''}",
+            f"memberships:{','.join(sorted(self.memberships))}",
+            f"photo_url:{self.photo_url or ''}",
         ]
 
         content_string = "\n".join(content_parts)
